+++ test/e2e/testing-manifests/storage-csi/mock/csi-mock-driver.yaml
@@ -44,6 +44,8 @@ spec:
                   apiVersion: v1
                   fieldPath: spec.nodeName
           imagePullPolicy: Always
+          securityContext:
+            privileged: true
           volumeMounts:
           - mountPath: /csi
             name: socket-dir
